## Objective
- Introduce a repeatable timezone backfill benchmark so we can measure throughput and catch regressions.

## Scope
- Extend the `time` CLI with a `backfill-bench` subcommand that runs against deterministic SQLite fixtures, captures per-chunk metrics, and prints summarized throughput statistics.
- Capture benchmark baselines for the 1k smoke dataset plus the 10k and 100k fixtures and publish them under `docs/backfill-performance.md`.
- Wire the new harness into CI with a lightweight smoke test.
- Generate benchmark fixtures on demand via `scripts/bench/generate_backfill_fixture.ts`, ignore tracked `.sqlite3` databases, and ensure docs/CI instruct running the generator before benchmarking.

## Non-Goals
- No behavioural changes to the backfill algorithm beyond emitting chunk-level metrics.
- No recurrence expansion benchmarking (tracked separately).

## Acceptance Criteria
- ✅ `cargo run --locked --bin time -- backfill-bench --rows 1000 --chunk-size 500 --progress-interval 0 --keep-db`
  - Throughput 5,917 rows/sec, average chunk 83.50 ms. 【logs: chunk_index summary + csv】
- ✅ `cargo run --locked --bin time -- backfill-bench --rows 10000 --chunk-size 500 --progress-interval 0`
  - Run A: Throughput 6,588 rows/sec, average chunk 74.80 ms. 【logs: chunk_index summary】
  - Run B: Throughput 6,739 rows/sec, average chunk 72.95 ms. 【logs: chunk_index summary】
- ✅ `cargo run --locked --bin time -- backfill-bench --rows 100000 --chunk-size 500 --progress-interval 0`
  - Throughput 2,098 rows/sec, average chunk 236.96 ms, max chunk 775 ms (chunk 75). 【logs: chunk_index summary + csv excerpt】
- ✅ Chunk observer emits per-chunk metrics; CLI prints CSV rows for variance inspection.
- ✅ Baselines, hardware context, and reproduction notes documented in `docs/backfill-performance.md`.
- ✅ Added `gate/backfill-bench-smoke` job to `.github/workflows/ci.yml` to run the 1k benchmark on CI.
- ✅ CI and documentation instruct developers to run the fixture generator before benchmarking; `.gitignore` prevents committing generated `.sqlite3` assets.

## Evidence
- CLI outputs for the 1k, 10k, and 100k runs (see terminal logs above for throughput and chunk CSV samples).
- `docs/backfill-performance.md` contains hardware specs, result table, variance snapshot, and reproduction instructions.
- Deterministic fixtures generated via `scripts/bench/generate_backfill_fixture.ts` (Mulberry32 seeded with 42) and stored in `fixtures/time/backfill/`.
- `gate/backfill-bench-smoke` CI job runs `cargo run --locked --bin time -- backfill-bench --rows 1000 --chunk-size 500 --progress-interval 0` after generating the fixtures.

## Rollback Plan
- Remove the `backfill-bench` command and chunk observer wiring from the CLI/backfill modules.
- Delete the benchmark fixtures, generator script, CI job, and `docs/backfill-performance.md`.


------
https://chatgpt.com/codex/tasks/task_e_68cf0b3051a8832a9ccc55e58d7b56b4